Allow globs for groups in configuration.
I have many projects written in a language, all of them in the same folder. I would like to run language-specific tools on each of the projects' files, but not configure the editor for each project separately. For that I need to be allowed to write a glob as a configuration group instead of a path.